Fix interference between reinforcement and capture
Before this change, it was possible for reinforcement move
to interfere with waiting capture plan.
Now capture plans take precedence, although it's done
in a dirty way: by overwriting reinforcement orders with
waiting capture orders.
Ideally it should be the other way around: reinforcement
planner should be aware which pieces it's not allowed to touch.